# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "zhuzhao——Go 模块化单体 IAM + 工单系统（三层鉴权 + 资源级工单）",
        "title": "zhuzhao API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/audit/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "审计日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "请求路径过滤",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期（YYYY-MM-DD）",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期（YYYY-MM-DD）",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工号",
                        "name": "employee_no",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回 Token 对",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户登出",
                "parameters": [
                    {
                        "description": "登出请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "刷新 Token",
                "parameters": [
                    {
                        "description": "刷新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "菜单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.LoginRequest": {
            "type": "object",
            "required": [
                "employee_no",
                "password"
            ],
            "properties": {
                "device_id": {
                    "description": "D2-01：max=64——device_id 进 RT JWT claims 与 Redis 键（refresh:{uid}:{device}）",
                    "type": "string",
                    "maxLength": 64
                },
                "device_name": {
                    "type": "string",
                    "maxLength": 64
                },
                "employee_no": {
                    "description": "D2-01：max 对齐 DB varchar(50)——employee_no 直接进 Redis 锁定键\n（lock:login:{employee_no}），无上限时可灌入超大键撑爆 maxmemory 驱逐安全键",
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.LogoutRequest": {
            "type": "object",
            "properties": {
                "device_id": {
                    "description": "与登录时一致；空则使用 default",
                    "type": "string"
                }
            }
        },
...
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "device_id": {
                    "description": "与登录时一致；空则使用 default",
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    }
}